- Updated log actions to show up only for packets they are built for
git-svn-id: https://svn.code.sf.net/p/dolserver/code/PacketLogConverter/trunk@1011 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ LogPacketAttribute.cs
@@ -8,17 +8,17 @@ namespace PacketLogConverter
 	[Serializable, AttributeUsage(AttributeTargets.Class, AllowMultiple = false)]
 	public class LogPacketAttribute : Attribute
 	{
-		private byte m_packetCode;
+		private int m_packetCode;
 		private float m_version;
 		private ePacketDirection m_direction;
 		private string m_description;
 
-		public LogPacketAttribute(byte packetCode, float version, ePacketDirection direction)
+		public LogPacketAttribute(int packetCode, float version, ePacketDirection direction)
 			: this(packetCode, version, direction, null)
 		{
 		}
 
-		public LogPacketAttribute(byte packetCode, float version, ePacketDirection direction, string packetDescription)
+		public LogPacketAttribute(int packetCode, float version, ePacketDirection direction, string packetDescription)
 		{
 			m_packetCode = packetCode;
 			m_version = version;
